+++ src/index.js
@@ -1,10 +1,15 @@
+import feedCataDog from '../src/feed-catadog.js';
+
 const donutButtons = document.querySelectorAll('.donut');
+
+const caterpillarDog = document.getElementById('caterpillar-dog');
 
 for (let i = 0; i < donutButtons.length; i++) {
     const donutButton = donutButtons[i];
 
     donutButton.addEventListener('click', () => {
-        console.log(donutButton);
+        feedCataDog(caterpillarDog, donutButton.value);
 
     }); 
 }
+
